Make information_schema tests reusable
`information_schema` needs to be tested on a per-connector basis. This
commit moves existing connector-specific information schema tests to
shared superclass so that they are run for all connectors.
